A few FrontMatter() tests, and fixes
Fix1: parseOpening() was just re-buffering excess bytes to be dealt with
on the next Parse() call. However, if the chunk contained the entirety
of the frontmatter, incoming chunks (if any) wouldn't be scanned by
parseClosing(). Instead, parseClosing() should have been called by
parseOpening on excess data.

Fix2: parseClosing() was marking `p.stage = notSeeking` on defer.
Meaning that parseClosing wouldn't be able to buffer continued chunks
while it looked for the closing pair.